+++ site/src/pages/sitemap.xml.ts
@@ -20,6 +20,7 @@ export const GET: APIRoute = () => {
     { loc: absoluteUrl('/vyhodnye/'), lastmod: now },
     { loc: absoluteUrl('/vystavki/'), lastmod: now },
     { loc: absoluteUrl('/festivali/'), lastmod: now },
+    { loc: absoluteUrl('/artefakty/'), lastmod: now },
     { loc: absoluteUrl('/populyarnoe/'), lastmod: now },
     { loc: absoluteUrl('/poisk/'), lastmod: now },
     { loc: absoluteUrl('/partners/'), lastmod: now },